+++ src/genstudio/js/api.js
@@ -1,4 +1,4 @@
-import { $StateContext, AUTOGRID_MIN as AUTOGRID_MIN_WIDTH, WidthContext } from "./context";
+import { $StateContext, AUTOGRID_MIN as AUTOGRID_MIN_WIDTH } from "./context";
 import { MarkSpec, PlotSpec } from "./plot";
 import { html } from "./utils";
 
@@ -10,13 +10,47 @@ import * as mobxReact from "mobx-react-lite";
 import * as React from "react";
 import * as ReactDOM from "react-dom/client";
 import * as render from "./plot/render";
-import { tw } from "./utils";
-
-const { useState, useEffect, useContext, useMemo, useCallback } = React
+import { tw, useContainerWidth } from "./utils";
+const { useState, useEffect, useContext, useMemo, useRef, useCallback } = React
+import Katex from "katex";
+import markdownItKatex from "./markdown-it-katex";
 
 export { render };
-const DEFAULT_GRID_GAP = "10px"
 export const CONTAINER_PADDING = 10;
+const KATEX_CSS_URL = "https://cdn.jsdelivr.net/npm/katex@0.16.11/dist/katex.min.css"
+
+window.katexCssLoaded = false;
+function loadKatexCss() {
+    if (window.katexCssLoaded) return;
+    window.katexCssLoaded = true;
+    const link = document.createElement('link');
+    link.rel = 'stylesheet';
+    link.href = KATEX_CSS_URL;
+    document.head.appendChild(link);
+
+}
+
+export function katex(tex) {
+    const containerRef = useRef(null);
+
+    useEffect(() => {
+        loadKatexCss();
+    }, []);
+
+    useEffect(() => {
+        if (containerRef.current) {
+            try {
+                Katex.render(tex, containerRef.current, {
+                    throwOnError: false
+                });
+            } catch (error) {
+                console.error('Error rendering KaTeX:', error);
+            }
+        }
+    }, [tex]);
+
+    return html`<div ref=${containerRef} />`;
+}
 
 const MarkdownItInstance = new MarkdownIt({
     html: true,
@@ -24,7 +58,13 @@ const MarkdownItInstance = new MarkdownIt({
     typographer: true
 });
 
+MarkdownItInstance.use(markdownItKatex)
+
 export function md(text) {
+    useEffect(() => {
+        loadKatexCss();
+    }, []);
+
     return html`<div className=${tw("prose")} dangerouslySetInnerHTML=${{ __html: MarkdownItInstance.render(text) }} />`;
 }
 
@@ -55,7 +95,6 @@ export const Slider = mobxReact.observer(
         step = step || 1;
 
         const $state = useContext($StateContext);
-        const availableWidth = useContext(WidthContext);
         const isAnimated = typeof fps === 'number' && fps > 0;
         const [isPlaying, setIsPlaying] = useState(isAnimated);
 
@@ -91,7 +130,7 @@ export const Slider = mobxReact.observer(
         const togglePlayPause = useCallback(() => setIsPlaying((prev) => !prev), []);
         if (options.visible !== true) return;
         return html`
-        <div className=${tw("text-base flex flex-col my-2 gap-2")} style=${{ width: availableWidth }}>
+        <div className=${tw("text-base flex flex-col my-2 gap-2 w-full")}>
           <div className=${tw("flex items-center justify-between")}>
             <span className=${tw("flex gap-2")}>
               <label>${label}</label>
@@ -177,53 +216,47 @@ export function repeat(data) {
 }
 export { d3, MarkSpec, Plot, PlotSpec, React, ReactDOM };
 
-export function Grid({ children, style, minWidth = AUTOGRID_MIN_WIDTH, gap = DEFAULT_GRID_GAP, aspectRatio = 1 }) {
-    const availableWidth = useContext(WidthContext);
-    const effectiveMinWidth = Math.min(minWidth, availableWidth);
+export function Grid({ children, style, minWidth = AUTOGRID_MIN_WIDTH, gap = 2, aspectRatio = 1 }) {
+    const [containerRef, containerWidth] = useContainerWidth();
+
+    const effectiveMinWidth = Math.min(minWidth, containerWidth);
     const gapSize = parseInt(gap);
 
-    const numColumns = Math.max(1, Math.min(Math.floor(availableWidth / effectiveMinWidth), children.length));
-    const itemWidth = (availableWidth - (numColumns - 1) * gapSize) / numColumns;
+    const numColumns = Math.max(1, Math.min(Math.floor(containerWidth / effectiveMinWidth), children.length));
+    const itemWidth = (containerWidth - (numColumns - 1) * gapSize) / numColumns;
     const itemHeight = itemWidth / aspectRatio;
     const numRows = Math.ceil(children.length / numColumns);
     const layoutHeight = numRows * itemHeight + (numRows - 1) * gapSize;
 
     const containerStyle = {
         display: 'grid',
-        gap,
         gridTemplateColumns: `repeat(${numColumns}, 1fr)`,
         gridAutoRows: `${itemHeight}px`,
-        height: `${layoutHeight}px`,
-        width: `${availableWidth}px`,
+        width: '100%',
         overflowX: 'auto',
         ...style
     };
 
     return html`
-    <${WidthContext.Provider} value=${itemWidth}>
-      <div style=${containerStyle}>
+    <div ref=${containerRef} class=${tw(`gap-${gap}`)} style=${containerStyle}>
         ${children.map((value, index) => html`<${Node} key=${index}
                                                        style=${{ width: itemWidth }}
                                                        value=${value}/>`)}
       </div>
-    </>
   `;
 }
 
 export function Row({ children, ...props }) {
-    const availableWidth = useContext(WidthContext);
-    const childCount = React.Children.count(children);
-    const childWidth = availableWidth / childCount;
+    const className = `flex flex-row ${props.className || ''}`
+    delete props["className"]
 
     return html`
-    <div ...${props} className=${tw("flex flex-row")}>
-      <${WidthContext.Provider} value=${childWidth}>
-        ${React.Children.map(children, (child, index) => html`
-          <div className=${tw("flex-1")} key=${index}>
-            ${child}
-          </div>
-        `)}
-      </${WidthContext.Provider}>
+    <div ...${props} className=${tw(className)}>
+      ${React.Children.map(children, (child, index) => html`
+        <div className=${tw("flex-1")} key=${index}>
+          ${child}
+        </div>
+      `)}
     </div>
   `;
 }